Finished passCards
Almost done with playTrick, I'm just having a little difficulty getting the list to recognize when there's only one element left with some of its methods. After that all I have left is just the scoring (which is super easy) and if I have time I'll implement a computer player. Game will be done after that.